requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
sqlalchemy==2.0.38
python-dotenv==1.2.1
